DateField: simplification
Validation is complicated: I faced questions like:

What do I do when an invalid date is passed into the value prop? Do I
replace it with a benign value? Do I throw an error? Do I show a
validation error next to the field? etc.

Neither of these felt perfect so I decided to just not let in invalid
values. Give the user ways to enter the date correctly and not worry
about its correctness. Just accept it as is.

After I decided this, I thought it looks like that old UNIXy advice:
“tools, not polices” — very cool.