Section component: wrote tests in preparation for refactoring
One thing that makes me smile while doing TDD is finding code that is
not needed. I think it’s because it makes me focus on every tiny bit and
question: Why are you here?

Tests also come in handy to document all these findings. This is good to
answer later questions like: Why did I do it this way?

“Test-after” seems to be a good code review technique.