Give up on my React component definition style
I wanted it so that I can have component’s definition members (like
propTypes, mixins, the render function) defined in the order that
reflects their importance: render ar the top, mixins at the bottom, etc.

I could do this if I defined the definition object as a separate plain
JS object — so now I can define members in multiple assignments across
the file. For example:

    var TextField = {};

    TextField.render = function() {...}

    ....

    TextField.mixins = [...];

and then call React.createClass(TextField) at the bottom to export it.

The standard definition style:

    var TextField = React.createClass({...});

is constraining in that it requires that the values in the definition
members are defined, and when they are not functions their definition
has to come before the components, which is noisy and makes the code
harder to scan and make sense of.

But my style doesn’t work well with private components. If I want a
small component that is used here and I want it to be in this file for
cohesion, it doesn’t work because I need a variable to hold the
component definition, and one other for the component itself:

    var CloseButton = {...};
    var CloseButtonItself = React.createClass(CloseButton);

which doesn’t work and I would have to fall back to the standard style,
which would result in two component styles, which I don’t like.

This is why I give up on my component definition style and solve the
definition members order issue in other ways.