DateField: prevent default on picker button click
An interesting bug. :o)

I found a whole bunch of test failing in Firefox and Safari. But there
was on error DateFieldTest that didn’t make sense: after clicking the
date picker button, the date-picker was not being found, only in Firefox
and Chrome. In Firefox it went away after updating it.

It turned out that even though the event on the picker button was being
stopped from propagation, it’s parent <label> was hearing the clicks
anyway - only in Firefox and Safari - and so the click was bubbling up
to the body and causing the date picker to be removed by the click
handler on <body> (I wanted the date picker to be hidden when the user
clicks anywhere else).

To get around that, I also had to prevent the default action of the
click.

Tyring to write the test for this, showed me that the global event
handlers (click on body and pressing Escape) actually belong to
DateField and not to DatePicker. Having that code in DatePicker, and
testing it’s side-effects in DateField test didn’t feel right, because
DatePicker is only a little nicety of DateField, and it shouldn’t care
about those details. It was like one half of event binding logic was in
one place, in the other half in some other place. Sincee DatePicker is
part of DateField, it can be argued that both halves were actually in
DatePicker, but event then, the two halves were not at the same level of
abstraction.

Interesting. It’s one of those noce side effects of testing: it creates
a perspective where you see things. :-)