Widgets.Section: accept a widget attribute
So far Section is used as a child DOM element by PersonSection,
SentenceSection, etc. Those widget create their own DOM element and have
Section’s DOM element as a child.

I want to be able to use Section’s DOM element as the SomeOtherSection
DOM element and keep the markup leaner, and, because it’s a <FIELDSET>,
I want to take advantage of its built-in semantics.

For that I need a way to differentiate between different types of
Section, and I’ll use the “widget” attribute for that.