TodoItem: getData() returns the id an label too
I think it’s a good idea to persist the label because it may change during the
application lifetime, or the todo item can be deleted/added. In that case I
want to still be able to render them nicely.

Because the id and label are passed into constructor as identifying
information, I think setData should ignore them.